Set verbosity level on root logger
Before the splitting off of car components into a dedicated package,
we only had a single logger, and hence setting the level on this
logger covered any relevant logging. After the splitting off, each car
component now comes with its own logger, so we need to cover all of
them. Instead of passing down the verbosity flag to each of them, we
switch to setting it on the root logger.